# storage: fix data race in MVCCPredicateDeleteRange