--- a/pkg/storage/testdata/mvcc_histories/delete_range_predicate
+++ b/pkg/storage/testdata/mvcc_histories/delete_range_predicate
@@ -47,7 +47,7 @@ meta: "i"/0,0 -> txn={id=00000001 key=/Min iso=Serializable pri=0.00000000 epo=0
 data: "i"/7.000000000,0 -> /BYTES/i7
 
 # Writing next to or above point keys and tombstones should work.
-run stats ok
+run stats ok log-ops
 del_range_pred k=a end=i ts=5 startTime=3 rangeThreshold=2
 ----
 >> del_range_pred k=a end=i ts=5 startTime=3 rangeThreshold=2
@@ -68,6 +68,8 @@ data: "g"/2.000000000,0 -> /BYTES/g2
 data: "h"/4.000000000,0 -> /BYTES/h4
 meta: "i"/0,0 -> txn={id=00000001 key=/Min iso=Serializable pri=0.00000000 epo=0 ts=7.000000000,0 min=0,0 seq=0} ts=7.000000000,0 del=false klen=12 vlen=7 mergeTs=<nil> txnDidNotUpdateMeta=true
 data: "i"/7.000000000,0 -> /BYTES/i7
+logical op: write_value: key="d", ts=5.000000000,0
+logical op: delete_range: startKey="f" endKey="h\x00" ts=5.000000000,0
 stats: key_count=8 key_bytes=160 val_count=12 val_bytes=111 range_key_count=2 range_key_bytes=27 range_val_count=2 live_count=3 live_bytes=111 gc_bytes_age=17863 intent_count=1 intent_bytes=19 lock_count=1 lock_age=93
 
 # Error on intent, no tombstones should be written. We try both the
@@ -228,7 +230,7 @@ stats: key_count=11 key_bytes=202 val_count=15 val_bytes=132 range_key_count=2 r
 
 # try the same call as above, except with startTime set to 1
 # check that delrange properly continues the run over a range tombstone
-run stats ok
+run stats ok log-ops
 del_range_pred k=j end=r ts=5 startTime=1 rangeThreshold=2
 ----
 >> del_range_pred k=j end=r ts=5 startTime=1 rangeThreshold=2
@@ -254,6 +256,7 @@ meta: "i"/0,0 -> txn={id=00000001 key=/Min iso=Serializable pri=0.00000000 epo=0
 data: "i"/7.000000000,0 -> /BYTES/i7
 data: "j"/2.000000000,0 -> /BYTES/j2
 data: "q"/2.000000000,0 -> /BYTES/q2
+logical op: delete_range: startKey="j" endKey="q\x00" ts=5.000000000,0
 stats: key_count=11 key_bytes=202 val_count=15 val_bytes=132 range_key_count=4 range_key_bytes=63 range_val_count=5 live_count=4 live_bytes=132 gc_bytes_age=25269 intent_count=1 intent_bytes=19 lock_count=1 lock_age=93
 
 # check that we flush with a range tombstone, if maxBytes is exceeded