@stevendanna 2023-10-30
@msbutler I played a bit with different allocation strategies but so far they've mostly been a mixed bag in your benchmark.  This change does create a big increase in allocations:

```
Without this change, we see a pretty large increase in the number of
allocations in our curren benchmark:

                                                                            │ master.txt  │    change_without_bufalloc.txt     │
                                                                            │   sec/op    │   sec/op     vs base               │
MVCCDeleteRangeWithPredicate_Pebble/streakBound=10/rangeKeyThreshold=64-10     4.144 ± 2%    4.182 ± 2%       ~ (p=0.221 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=100/rangeKeyThreshold=64-10    1.808 ± 1%    1.791 ± 1%       ~ (p=0.211 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=200/rangeKeyThreshold=64-10   574.4m ± 1%   594.8m ± 3%  +3.55% (p=0.000 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=500/rangeKeyThreshold=64-10   60.50m ± 3%   62.94m ± 2%  +4.03% (p=0.000 n=20)
geomean                                                                       714.3m        727.6m       +1.87%

                                                                            │ master.txt  │     change_without_bufalloc.txt     │
                                                                            │  allocs/op  │  allocs/op   vs base                │
MVCCDeleteRangeWithPredicate_Pebble/streakBound=10/rangeKeyThreshold=64-10    2.188k ± 1%   3.424k ± 1%  +56.47% (p=0.000 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=100/rangeKeyThreshold=64-10   1.695k ± 2%   2.518k ± 1%  +48.58% (p=0.000 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=200/rangeKeyThreshold=64-10   1.668k ± 1%   2.026k ± 2%  +21.46% (p=0.000 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=500/rangeKeyThreshold=64-10   2.026k ± 1%   2.163k ± 1%   +6.71% (p=0.000 n=20)
geomean                                                                       1.882k        2.479k       +31.75%
```

Using our `bufalloc` package, I can bring the allocation cost down substantially, but at the cost of the runtime of the benchmark:


```
                                                                            │ master.txt  │      change_with_bufalloc.txt      │
                                                                            │   sec/op    │   sec/op     vs base               │
MVCCDeleteRangeWithPredicate_Pebble/streakBound=10/rangeKeyThreshold=64-10     4.144 ± 2%    4.437 ± 3%  +7.08% (p=0.000 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=100/rangeKeyThreshold=64-10    1.808 ± 1%    1.921 ± 2%  +6.27% (p=0.000 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=200/rangeKeyThreshold=64-10   574.4m ± 1%   607.2m ± 2%  +5.72% (p=0.000 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=500/rangeKeyThreshold=64-10   60.50m ± 3%   65.47m ± 4%  +8.20% (p=0.000 n=20)
geomean                                                                       714.3m        763.0m       +6.81%

                                                                            │ master.txt  │      change_with_bufalloc.txt       │
                                                                            │  allocs/op  │  allocs/op   vs base                │
MVCCDeleteRangeWithPredicate_Pebble/streakBound=10/rangeKeyThreshold=64-10    2.188k ± 1%   2.538k ± 1%  +15.97% (p=0.000 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=100/rangeKeyThreshold=64-10   1.695k ± 2%   1.687k ± 3%        ~ (p=0.189 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=200/rangeKeyThreshold=64-10   1.668k ± 1%   1.639k ± 1%   -1.77% (p=0.002 n=20)
MVCCDeleteRangeWithPredicate_Pebble/streakBound=500/rangeKeyThreshold=64-10   2.026k ± 1%   1.970k ± 2%   -2.76% (p=0.000 n=20)
geomean                                                                       1.882k        1.928k        +2.47%
```

Perhaps this is worth the trade-off in the real world.